+++ react/components/ProductList.js
@@ -37,14 +37,19 @@ const ProductList = ({
   minItemsPerPage,
   paginationDotsVisibility,
   navigationStep: navigationStepProp,
+  trackingId,
 }) => {
   const handles = useCssHandles(CSS_HANDLES)
   const navigationStep = Number.isNaN(parseInt(navigationStepProp, 10))
     ? navigationStepProp
     : parseInt(navigationStepProp, 10)
 
-  const formattedTitle =
-    showTitle && titleText ? formatIOMessage(titleText) : 'List of products'
+  let listName = trackingId
+
+  if (!listName) {
+    listName =
+      showTitle && titleText ? formatIOMessage(titleText) : 'List of products'
+  }
 
   return products && !products.length ? null : (
     <Fragment>
@@ -71,7 +76,7 @@ const ProductList = ({
             navigationStep={navigationStep}
             minItemsPerPage={minItemsPerPage}
             paginationDotsVisibility={paginationDotsVisibility}
-            listName={formattedTitle}
+            listName={listName}
           />
         )}
       </ReactResizeDetector>
@@ -106,6 +111,7 @@ ProductList.propTypes = {
     'desktopOnly',
     'mobileOnly',
   ]),
+  trackingId: PropTypes.string,
   ...productListSchemaPropTypes,
 }